refactoring for tests and added tests
