+++ TechTalk.SpecFlow.Generator/DefaultDependencyProvider.cs
@@ -3,6 +3,7 @@ using TechTalk.SpecFlow.Generator.Configuration;
 using TechTalk.SpecFlow.Generator.Interfaces;
 using TechTalk.SpecFlow.Generator.Plugins;
 using TechTalk.SpecFlow.Generator.UnitTestConverter;
+using TechTalk.SpecFlow.PlatformSpecific;
 using TechTalk.SpecFlow.Utils;
 
 namespace TechTalk.SpecFlow.Generator
@@ -32,6 +33,8 @@ namespace TechTalk.SpecFlow.Generator
 
             container.RegisterInstanceAs(GenerationTargetLanguage.CreateCodeDomHelper(GenerationTargetLanguage.CSharp), GenerationTargetLanguage.CSharp, dispose: true);
             container.RegisterInstanceAs(GenerationTargetLanguage.CreateCodeDomHelper(GenerationTargetLanguage.VB), GenerationTargetLanguage.VB, dispose: true);
+
+            container.RegisterTypeAs<RuntimeConfigurationLoader, IRuntimeConfigurationLoader>();
 
             RegisterUnitTestGeneratorProviders(container);
         }
