+++ TechTalk.SpecFlow.Reporting/NUnitExecutionReport/NUnitBasedExecutionReportGenerator.cs
@@ -4,6 +4,7 @@ using System.Xml;
 using System.Xml.Serialization;
 using TechTalk.SpecFlow.Generator.Configuration;
 using TechTalk.SpecFlow.Generator.Project;
+using TechTalk.SpecFlow.Tracing;
 
 namespace TechTalk.SpecFlow.Reporting.NUnitExecutionReport
 {
@@ -35,7 +36,7 @@ namespace TechTalk.SpecFlow.Reporting.NUnitExecutionReport
 
         public void GenerateAndTransformReport()
         {
-            var specFlowProject = MsBuildProjectReader.LoadSpecFlowProjectFromMsBuild(ReportParameters.ProjectFile);            
+            var specFlowProject = MsBuildProjectReader.LoadSpecFlowProjectFromMsBuild(ReportParameters.ProjectFile, new NullListener());            
 
             var report = GenerateReport(specFlowProject);
             TransformReport(report, specFlowProject);
