+++ Tests/TechTalk.SpecFlow.RuntimeTests/Configuration/AppConfigTests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Linq;
 using BoDi;
 using FluentAssertions;
@@ -6,7 +7,9 @@ using NUnit.Framework;
 using TechTalk.SpecFlow.BindingSkeletons;
 using TechTalk.SpecFlow.Configuration;
 using TechTalk.SpecFlow.PlatformSpecific;
+using TechTalk.SpecFlow.PlatformSpecific.AppConfig;
 using TechTalk.SpecFlow.Plugins;
+using TechTalk.SpecFlow.Tracing;
 
 namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 {
@@ -17,7 +20,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
         public void CanLoadConfigFromConfigFile()
         {
             var runtimeConfiguration = RuntimeConfigurationLoader.GetDefault();
-            var configurationLoader = new RuntimeConfigurationLoader();
+            var configurationLoader = new RuntimeConfigurationLoader(new TextWriterTraceListener(new StringWriter()));
 
             runtimeConfiguration = configurationLoader.Load(runtimeConfiguration);
         }
@@ -46,7 +49,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
         {
             var runtimeConfig = RuntimeConfigurationLoader.GetDefault();
 
-            var configurationLoader = new RuntimeConfigurationLoader();
+            var configurationLoader = new AppConfigConfigurationLoader();
 
 
             var configurationSectionHandler = ConfigurationSectionHandler.CreateFromXml(configString);
@@ -60,7 +63,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.FeatureLanguage.TwoLetterISOLanguageName.Should().Be("de");
         }
@@ -72,7 +75,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.ToolLanguage.TwoLetterISOLanguageName.Should().Be("de");
         }
@@ -84,7 +87,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.BindingCulture.TwoLetterISOLanguageName.Should().Be("de");
         }
@@ -96,7 +99,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.RuntimeUnitTestProvider.Should().Be("XUnit");
         }
@@ -108,7 +111,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.DetectAmbiguousMatches.Should().BeTrue();
         }
@@ -120,7 +123,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.DetectAmbiguousMatches.Should().BeFalse();
         }
@@ -132,7 +135,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.StopAtFirstError.Should().BeTrue();
         }
@@ -144,7 +147,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.StopAtFirstError.Should().BeFalse();
         }
@@ -156,7 +159,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.MissingOrPendingStepsOutcome.Should().Be(MissingOrPendingStepsOutcome.Pending);
         }
@@ -168,7 +171,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.MissingOrPendingStepsOutcome.Should().Be(MissingOrPendingStepsOutcome.Error);
         }
@@ -180,7 +183,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.MissingOrPendingStepsOutcome.Should().Be(MissingOrPendingStepsOutcome.Ignore);
         }
@@ -192,7 +195,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.MissingOrPendingStepsOutcome.Should().Be(MissingOrPendingStepsOutcome.Inconclusive);
         }
@@ -204,7 +207,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.TraceSuccessfulSteps.Should().BeTrue();
         }
@@ -216,7 +219,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.TraceSuccessfulSteps.Should().BeFalse();
         }
@@ -228,7 +231,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.TraceTimings.Should().BeTrue();
         }
@@ -240,7 +243,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.TraceTimings.Should().BeFalse();
         }
@@ -252,7 +255,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.MinTracedDuration.Should().Be(TimeSpan.FromSeconds(1));
         }
@@ -264,7 +267,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.CustomDependencies.Count.Should().Be(1);
 
@@ -281,7 +284,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.StepDefinitionSkeletonStyle.Should().Be(StepDefinitionSkeletonStyle.RegexAttribute);
         }
@@ -293,7 +296,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.StepDefinitionSkeletonStyle.Should().Be(StepDefinitionSkeletonStyle.MethodNamePascalCase);
         }
@@ -305,7 +308,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.StepDefinitionSkeletonStyle.Should().Be(StepDefinitionSkeletonStyle.MethodNameRegex);
         }
@@ -317,7 +320,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.StepDefinitionSkeletonStyle.Should().Be(StepDefinitionSkeletonStyle.MethodNameUnderscores);
         }
@@ -329,7 +332,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.AdditionalStepAssemblies.Should().BeEmpty();
         }
@@ -341,7 +344,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.AdditionalStepAssemblies.Should().BeEmpty();
         }
@@ -353,7 +356,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.AdditionalStepAssemblies.Count.Should().Be(1);
             runtimeConfig.AdditionalStepAssemblies.First().Should().Be("testEntry");
@@ -369,7 +372,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.AdditionalStepAssemblies.Count.Should().Be(2);
             runtimeConfig.AdditionalStepAssemblies[0].Should().Be("testEntry1");
@@ -383,7 +386,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Should().BeEmpty();
         }
@@ -395,7 +398,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Should().BeEmpty();
         }
@@ -407,7 +410,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(1);
             runtimeConfig.Plugins.First().Name.Should().Be("testEntry");
@@ -420,7 +423,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(1);
             runtimeConfig.Plugins.First().Path.Should().Be("path_to_assembly");
@@ -433,7 +436,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(1);
             runtimeConfig.Plugins.First().Parameters.Should().Be("pluginParameter");
@@ -446,7 +449,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(1);
             runtimeConfig.Plugins.First().Type.Should().Be(PluginType.Runtime);
@@ -459,7 +462,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(1);
             runtimeConfig.Plugins.First().Type.Should().Be(PluginType.Generator);
@@ -472,7 +475,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(1);
             runtimeConfig.Plugins.First().Type.Should().Be(PluginType.GeneratorAndRuntime);
@@ -488,7 +491,7 @@ namespace TechTalk.SpecFlow.RuntimeTests.Configuration
 
             var configSection = ConfigurationSectionHandler.CreateFromXml(config);
 
-            var runtimeConfig = new RuntimeConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
+            var runtimeConfig = new AppConfigConfigurationLoader().LoadAppConfig(RuntimeConfigurationLoader.GetDefault(), configSection);
 
             runtimeConfig.Plugins.Count.Should().Be(2);
             runtimeConfig.Plugins[0].Name.Should().Be("testEntry1");
